Updated new testing SchemaTypes and the Structure of the TypeBaseClass
