+++ pressbooks-metadata/admin/schemaTypes/creativeWorks/class-pressbooks-metadata-blog.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace schemaTypes\cw;
-use schemaFunctions\Pressbooks_Metadata_Create_Metabox as create_metabox;
+use schemaTypes;
 use schemaTypes\Pressbooks_Metadata_Type;
 
 /**
@@ -24,7 +24,15 @@ class Pressbooks_Metadata_Blog extends Pressbooks_Metadata_Type {
 	 * @since    0.x
 	 * @access   public
 	 */
-	const type_setting = array('blog_type' => array('Blog Type','http://schema.org/Blog'));
+	static $type_setting = array('blog_type' => array('Blog Type','http://schema.org/Blog'));
+
+	/**
+	 * The variable that holds the parents for the type
+	 *
+	 * @since    0.x
+	 * @access   public
+	 */
+	static $type_parents = array('schemaTypes\Pressbooks_Metadata_Thing');
 
 	/**
 	 * The variable that holds the properties of this schema type
@@ -32,17 +40,30 @@ class Pressbooks_Metadata_Blog extends Pressbooks_Metadata_Type {
 	 * @since    0.x
 	 * @access   public
 	 */
-	const type_properties = array(
+	static $type_properties = array(
 		'blogPost' => array(true,'Blog Post','A posting that is part of this blog.')
 	);
 
 	public function __construct($type_level_input) {
 		parent::__construct($type_level_input);
-		$this->type_fields = self::type_properties;
+		$this->type_fields = $this->get_all_properties();
 		$this->class_name = __CLASS__ .'_'. $this->type_level;
-		//$this->parent_type = new Pressbooks_Metadata_Creative_Work($this->type_level);
-		$this->pmdt_populate_names(self::type_setting);
+		$this->pmdt_populate_names(self::$type_setting);
 		$this->pmdt_add_metabox($this->type_level);
+	}
+
+	/**
+	 * Function used for combining the current types properties with its parents fields
+	 *
+	 * @since    0.x
+	 * @access   public
+	 */
+	public function get_all_properties() {
+		$properties = self::$type_properties;
+		foreach(self::$type_parents as $parentType){
+			$properties = array_merge($properties,$parentType::type_properties);
+		}
+		return $properties;
 	}
 
 	/**
@@ -53,37 +74,5 @@ class Pressbooks_Metadata_Blog extends Pressbooks_Metadata_Type {
 	 */
 	public function __toString() {
 		return $this->class_name;
-	}
-
-	/**
-	 * The function which produces the metaboxes for the book type
-	 * @param string Accepting a string so we can distinguish on witch place each metabox is created
-	 * The value passed here is also used when calling the metadata functions in the header and the footer.
-	 * @since 0.8.1
-	 */
-	private function pmdt_add_metabox($meta_position) {
-		new create_metabox($this->typeName,$this->typeDisplayName,$meta_position,$this->type_fields,NULL);
-	}
-
-	/**
-	 * A function that creates the metadata for the book type.
-	 * @since 0.8.1
-	 *
-	 */
-	public function pmdt_get_metatags() {
-		//Creating microtags
-		$html = "<!-- Microtags --> \n";
-
-		$html .= '<div itemscope itemtype="http://schema.org/Blog">';
-
-		foreach ( $this->type_fields as $itemprop => $details ) {
-			$propName = strtolower('pb_' . $itemprop . '_' . $this->type_level);
-			if ($this->pmdt_prop_run($itemprop)) {
-				$value = $this->pmdt_get_value($propName);
-				if(!empty($value)){$html .= "<meta itemprop = '" . $itemprop . "' content = '" . $value . "'>\n";}
-			}
-		}
-		$html .= '</div>';
-		return $html;
 	}
 }
